architecture du back + probleme decriture du json rajout de ]} en trop dans le monde des user apres saveword

+++ backend/schema.js
@@ -10,7 +10,7 @@ module.exports = gql`
         logo: String
         seuil: Float
         idcible: Int
-        ratio: Float
+        ratio: Int
         typeratio: RatioType
         unlocked: Boolean
     }
